refactor(cloudflare-metrics): move handlers out of index.ts
index.ts was the worker entry point AND the implementation of both the
http fetch handler and the scheduled cron handler. split so:

- handlers/http.ts        handleFetch + routing + error-path boilerplate
- handlers/scheduled.ts   handleScheduled + runCollection + flush stats
- index.ts                just the default export wiring the two

the handler tests import from ./index.js so no test changes are needed.
index.ts drops from 174 to 19 lines.
